ecommerce_system/product_service/requirements.txt
# product_service/requirements.txt
Flask
pika
psycopg2-binary